Added deduplication of triangulations
In practice, this should **not** matter at all, but I want to ensure
that, when merging different datasets, no errors creep in. There are
also other factors to consider, viz., possible extensions for a more
elaborate way of *generating* triangulations. This will also require
a good method to find duplicates.